refactor: update jsdocs for utils, errors and user resources
