+++ src/utils/asyncWrapper.js
@@ -1,9 +1,10 @@
 /**
+ * Wrapper the try/catch into a promise
  * @module asyncWrapper
  */
 
 /**
- * @callback functionToWrap
+ * @callback cbTemplate
  * @param {{}} req - Request
  * @param {{}} res - Response
  * @param {function} [next]
@@ -13,8 +14,8 @@
 /**
  * Function to wrap the try/catch into a promise for any Express handler
  * @see https://zellwk.com/blog/async-await-express/
- * @param {functionToWrap} callback - a function to be wrapped  
- * @returns {function} function that wraps a given callback
+ * @param {cbTemplate} callback - a function to be wrapped  
+ * @returns {cbTemplate} function that wraps a given callback
  */
 function asyncWrapper(callback) {
   return (req, res, next) => {
